Generate site twice on deploy
Generate the first time so that the output directories are
created. Then when PurgeCSS runs, it can look at the generated
files when purging the stylesheets.

Generate again to copy the final assets from "source/build" to
the correct output directory.

There may be a better way, but this works for now.