Use Traefik as a local reverse proxy
Use Traefik as a local reverse proxy so that the site can be accessed
from http://oliverdavies.localhost rather than http://localhost:8000,
which is the default for Sculpin.

I've also overridden the command used for the app service so that it
exposes port 80 and serves the site on that port, as well as configuring
the URL to match within the generated output.