build(docker): make Compose file more configurable
Allow for configuring volume mounts, ports, the site URL, and volume
mounts within the Docker Compose file using environment variables from
within and .env file.

This means that the same Docker Compose file could be used to generate
and serve the site locally using an anonymous volume, as well as
generate a one-time build into a local directory to be served by Apache
or Nginx on production.