chore(nginx): simplify vhost configuration
Simplify the vhost configuration now that Caddy is sitting in front of
Nginx again.

This means that Caddy is responsible for the SSL certificate generation
and configuration, and proxying requests to Nginx, which is responsible
for processing the redirect map and serving the static files from disk.

This is a temporary solution until I can get the redirects working in
the same way with Caddy.