Only past talks are counted
Talks at an event happening "today" haven't happened yet, so shouldn't
be counted.

A talk should only counted if it was given at a past event -
i.e. yesterday or earlier.

Previously a talk with today's date was returning a time of 00:00 and
was being counted in the past count if the site was generated on the
same day.